pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "q-imgen"
version = "0.3.0"
description = "Atomic nanobanana CLI with channel-based endpoint routing"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
dependencies = ["Pillow>=10.0"]
authors = [
  { name = "ywt" }
]
keywords = ["image-generation", "cli", "automation", "gemini", "nanobanana"]
classifiers = [
  "Development Status :: 4 - Beta",
  "Intended Audience :: Developers",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Environment :: Console",
  "Topic :: Multimedia :: Graphics",
  "Topic :: Software Development :: Libraries :: Python Modules"
]

[project.scripts]
q-imgen = "q_imgen.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
q_imgen = []

[tool.pyright]
include = ["src", "tests"]
extraPaths = ["src"]
